requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi
uvicorn
sqlalchemy
pillow
torch
torchvision
python-dotenv
google-generativeai
ultralytics
opencv-python-headless
onnxruntime
numpy
python-multipart
